reorganized webapp structure + added api clients for the rest of the services
