+++ source/webapp/src/App.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import {BrowserRouter as Router, Routes, Route} from "react-router-dom";
-import Dashboard from "./pages/Dashboard";
-import Login from "./pages/Login";
-import Register from "./pages/Register";
-import EncryptDecrypt from "./pages/EncryptDecrypt";
-import EntriesVisualization from "./pages/EntriesVisualization";
-import Wiki from "./pages/Wiki";
+import Dashboard from "./pages/dashboard/Dashboard";
+import Login from "./pages/login/Login";
+import Register from "./pages/register/Register";
+import EncryptDecrypt from "./pages/encrypt-decrypt/EncryptDecrypt";
+import EntriesVisualization from "./pages/entries-visualization/EntriesVisualization";
+import Wiki from "./pages/wiki/Wiki";
 import {Provider} from "jotai";
 
 function App() {
