+++ source/webapp/src/components/navigation/navigation.css
@@ -47,7 +47,8 @@
     position: relative;
     text-decoration: none;
     font-weight: 500;
-    font-size: 18px;
+    font-size: 16px;
+    margin-top: 2px;
 }
 
 .sidebar-menu-nav-link {
