+++ source/webapp/src/pages/dashboard/Dashboard.tsx
@@ -1,13 +1,13 @@
 import {Col, Container, Row} from "react-bootstrap";
+import Navigation from "../../components/navigation/Navigation";
 import React from "react";
+import {userAtom} from "../../services/UserAtom";
 import {useAtom} from "jotai";
-import {userAtom} from "../services/UserAtom";
-import Header from "../components/header/Header";
-import Navigation from "../components/navigation/Navigation";
-import EncryptDecryptIcon from "../resources/icons/menu/EncryptDecryptIcon";
-import Loader from "../components/loader/Loader";
+import Header from "../../components/header/Header";
+import DashboardIcon from "../../utils/resources/icons/menu/DashboardIcon";
+import Loader from "../../components/loader/Loader";
 
-const EncryptDecrypt = () => {
+const Dashboard = () => {
 
     const [currentUser,] = useAtom(userAtom);
     currentUser === undefined && window.open("/login", "_self");
@@ -24,11 +24,11 @@ const EncryptDecrypt = () => {
                     <Container fluid>
                         <Row>
                             <Col xl={{span: 2}} lg={{span: 3}} className="no-padding-left no-padding-right">
-                                <Navigation currentPage="/encrypt-decrypt"/>
+                                <Navigation currentPage="/"/>
                             </Col>
                             <Col xl={{span: 10}} lg={{span: 9}} className="no-padding-right no-padding-left">
                                 <Row>
-                                    <Header icon={<EncryptDecryptIcon/>} text={"Encrypt / Decrypt"}/>
+                                    <Header icon={<DashboardIcon/>} text={"Dashboard"}/>
                                 </Row>
                             </Col>
                         </Row>
@@ -39,4 +39,4 @@ const EncryptDecrypt = () => {
     }
 }
 
-export default EncryptDecrypt;
+export default Dashboard;
