+++ source/webapp/src/pages/encrypt-decrypt/EncryptDecrypt.tsx
@@ -1,13 +1,13 @@
 import {Col, Container, Row} from "react-bootstrap";
-import Navigation from "../components/navigation/Navigation";
 import React from "react";
-import {userAtom} from "../services/UserAtom";
 import {useAtom} from "jotai";
-import Header from "../components/header/Header";
-import DashboardIcon from "../resources/icons/menu/DashboardIcon";
-import Loader from "../components/loader/Loader";
+import {userAtom} from "../../services/UserAtom";
+import Header from "../../components/header/Header";
+import Navigation from "../../components/navigation/Navigation";
+import EncryptDecryptIcon from "../../utils/resources/icons/menu/EncryptDecryptIcon";
+import Loader from "../../components/loader/Loader";
 
-const Dashboard = () => {
+const EncryptDecrypt = () => {
 
     const [currentUser,] = useAtom(userAtom);
     currentUser === undefined && window.open("/login", "_self");
@@ -24,11 +24,11 @@ const Dashboard = () => {
                     <Container fluid>
                         <Row>
                             <Col xl={{span: 2}} lg={{span: 3}} className="no-padding-left no-padding-right">
-                                <Navigation currentPage="/"/>
+                                <Navigation currentPage="/encrypt-decrypt"/>
                             </Col>
                             <Col xl={{span: 10}} lg={{span: 9}} className="no-padding-right no-padding-left">
                                 <Row>
-                                    <Header icon={<DashboardIcon/>} text={"Dashboard"}/>
+                                    <Header icon={<EncryptDecryptIcon/>} text={"Encrypt / Decrypt"}/>
                                 </Row>
                             </Col>
                         </Row>
@@ -39,4 +39,4 @@ const Dashboard = () => {
     }
 }
 
-export default Dashboard;
+export default EncryptDecrypt;
