+++ source/webapp/src/pages/entries-visualization/entriesVisualization.css
@@ -1,0 +1,4 @@
+.delete-entry-button {
+    display: flex;
+    justify-content: end;
+}
