+++ source/webapp/src/pages/register/Register.tsx
@@ -3,19 +3,19 @@ import {Formik, Field, Form, ErrorMessage} from "formik";
 import {Col, Container, Row} from "react-bootstrap";
 import * as Yup from "yup";
 
-import SmallTitle from "../components/typography/titles/SmallTitle";
-import GenericText from "../components/typography/text/GenericText";
-import Logo from "../resources/Logo";
+import SmallTitle from "../../components/typography/titles/SmallTitle";
+import GenericText from "../../components/typography/text/GenericText";
+import Logo from "../../utils/resources/Logo";
 
-import {RegisterCredentials, User} from "../apidocs/v1_pb";
-import {UserServiceApiClient} from "../api/UserServiceApiClient";
+import {RegisterCredentials, User} from "../../apidocs/v1_pb";
+import {UserServiceApiClient} from "../../services/api/UserServiceApiClient";
 import grpcWeb from "grpc-web";
 import {useAtom} from "jotai";
-import {userAtom} from "../services/UserAtom";
+import {userAtom} from "../../services/UserAtom";
 
-import "../helpers/forms.css";
-import "../helpers/styles.css";
-import "../helpers/paddings.css";
+import "../../utils/css/forms.css";
+import "../../utils/css/styles.css";
+import "../../utils/css/paddings.css";
 
 const Register = () => {
 
@@ -135,7 +135,7 @@ const Register = () => {
                                             {
                                                 message && (
                                                     <div className="form-group">
-                                                        <div className="alert alert-danger alert-with-background d-flex justify-content-center" role="alert">
+                                                        <div className="alert alert-danger d-flex justify-content-center" role="alert">
                                                             {message}
                                                         </div>
                                                     </div>
