+++ source/webapp/src/services/api/BATServiceApiClient.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+import {BATServiceClient} from "../../apidocs/v1_grpc_web_pb";
+
+export const BATServiceApiClient = new BATServiceClient('http://localhost:5000');
