+++ source/webapp/src/services/api/CTRUServiceApiClient.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+import {CTRUServiceClient} from "../../apidocs/v1_grpc_web_pb";
+
+export const CTRUServiceApiClient = new CTRUServiceClient('http://localhost:5000');
