+++ source/webapp/src/services/api/UserServiceApiClient.tsx
@@ -1,5 +1,5 @@
 import React from "react";
 
-import {UserServiceClient} from "../apidocs/v1_grpc_web_pb";
+import {UserServiceClient} from "../../apidocs/v1_grpc_web_pb";
 
 export const UserServiceApiClient = new UserServiceClient('http://localhost:5000');
