+++ source/webapp/src/utils/css/paddings.css
