+++ source/webapp/src/utils/resources/icons/menu/DashboardIcon.tsx
