+++ source/webapp/src/utils/resources/icons/menu/EncryptDecryptIcon.tsx
