+++ source/webapp/src/utils/resources/icons/menu/EntriesVisualizationIcon.tsx
