+++ source/webapp/src/utils/resources/icons/menu/LogoutIcon.tsx
